Replce ref option with filename option
Existing code always packaged HEAD regardless of ref passed in. Since
creating an artifact involves running assets:precompile, HEAD will
always be used as the package name.

Because it may be beneficial to have the ability to create the artifact
under a custom filename, that has been added as an option instead.

Change-Id: I1bc99ab23b892cfff5b0c414c5184a56e29c1374